+++ crypto.cpp
@@ -1,5 +1,6 @@
 #include <iostream>
 #include <cstdint>
+#include "DecodedInst.hpp"
 #include "Hart.hpp"
 #include "wideint.hpp"
 
@@ -970,7 +971,7 @@ Hart<URV>::execSm3p0(const DecodedInst* di)
     }
 
   uint32_t r1     = intRegs_.read(di->op1());
-  int32_t  sres32 = r1 ^ rol(r1, 9) ^ rol(r1, 17);
+  int32_t  sres32 = static_cast<int32_t>(r1 ^ rol(r1, 9) ^ rol(r1, 17));
   SRV      res    = sres32;  // sign extend.
 
   intRegs_.write(di->op0(), res);
@@ -988,7 +989,7 @@ Hart<URV>::execSm3p1(const DecodedInst* di)
     }
 
   uint32_t r1     = intRegs_.read(di->op1());
-  int32_t  sres32 = r1 ^ rol(r1, 15) ^ rol(r1, 23);
+  int32_t  sres32 = static_cast<int32_t>(r1 ^ rol(r1, 15) ^ rol(r1, 23));
   SRV      res    = sres32;  // sign extend.
 
   intRegs_.write(di->op0(), res);
